perf: :zap: reduced 20kB by isolating _app service calls
