+++ src/components/pages/document/common/use-document-filter.tsx
@@ -2,9 +2,9 @@ import useDidUpdateEffect from "@hooks/use-did-update-effect";
 import useGlobalState from "@hooks/use-global-state";
 import { DocumentData } from "@interfaces/custom";
 import { UserGroupIbp } from "@interfaces/document";
+import { axGroupList } from "@services/app.service";
 import { axGetDocumentSpeciesGroups, axGetListData } from "@services/document.service";
 import { axGetLandscapeList } from "@services/landscape.service";
-import { axGroupList } from "@services/usergroup.service";
 import { axGetAllHabitat } from "@services/utility.service";
 import { isBrowser } from "@static/constants";
 import { DEFAULT_FILTER, LIST_PAGINATION_LIMIT } from "@static/documnet-list";
