+++ src/pages/_app.tsx
@@ -2,7 +2,6 @@ import "../styles/global.css";
 
 import { ChakraProvider, createStandaloneToast } from "@chakra-ui/react";
 import { SkipNavContent, SkipNavLink } from "@chakra-ui/skip-nav";
-import AuthWall from "@components/@core/container/authwall";
 import Footer from "@components/@core/container/footer";
 import Metadata from "@components/@core/container/metadata";
 import NavigationMenuDark from "@components/@core/navigation-menu/dark";
@@ -11,14 +10,19 @@ import SITE_CONFIG from "@configs/site-config";
 import { customTheme } from "@configs/theme";
 import { GlobalStateProvider } from "@hooks/use-global-state";
 import { UserGroupIbp } from "@interfaces/observation";
-import { axGroupList } from "@services/usergroup.service";
+import { axGroupList } from "@services/app.service";
 import { getParsedUser } from "@utils/auth";
 import { absoluteUrl } from "@utils/basic";
 import App, { AppContext } from "next/app";
+import dynamic from "next/dynamic";
 import Router from "next/router";
 import NProgress from "nprogress";
 import React, { useEffect } from "react";
 import BusProvider from "react-gbus";
+
+const AuthWall = dynamic(() => import("@components/@core/container/authwall"), {
+  ssr: false
+});
 
 interface AppProps {
   Component;
@@ -28,9 +32,18 @@ interface AppProps {
   user;
   pageProps;
   pages;
+  languageId;
 }
 
-function MainApp({ Component, currentGroup, domain, groups, user, pageProps }: AppProps) {
+function MainApp({
+  Component,
+  currentGroup,
+  domain,
+  groups,
+  user,
+  languageId,
+  pageProps
+}: AppProps) {
   const { ToastContainer } = createStandaloneToast();
   const config = { header: true, footer: true, ...Component?.config };
 
@@ -44,7 +57,7 @@ function MainApp({ Component, currentGroup, domain, groups, user, pageProps }: A
     <BusProvider>
       <ToastContainer />
       <ChakraProvider theme={customTheme}>
-        <GlobalStateProvider initialState={{ user, domain, groups, currentGroup }}>
+        <GlobalStateProvider initialState={{ user, domain, groups, currentGroup, languageId }}>
           <Metadata />
           <SkipNavLink>Skip to content</SkipNavLink>
           <div className="content">
@@ -73,6 +86,7 @@ MainApp.getInitialProps = async (appContext: AppContext) => {
   const aReq = absoluteUrl(appContext.ctx, appContext.router.asPath);
   const domain = aReq.hostname.split(".").slice(-2).join(".");
   const user = getParsedUser(appContext.ctx);
+  const languageId = SITE_CONFIG.LANG.LIST[appContext.ctx.locale]?.ID;
 
   const { currentGroup, groups } = await axGroupList(aReq.href, appContext.ctx.locale);
 
@@ -81,7 +95,8 @@ MainApp.getInitialProps = async (appContext: AppContext) => {
     groups,
     currentGroup,
     domain,
-    user
+    user,
+    languageId
   };
 };
