+++ src/pages/datatable/list.tsx
@@ -1,9 +1,9 @@
 import { DataTableFilterContextProvider } from "@components/pages/datatable/common/use-datatable-filter";
 import { DEFAULT_PARAMS } from "@components/pages/datatable/common/use-datatableObservation-filter";
 import DataTableListPageComponent from "@components/pages/datatable/list";
+import { axGroupList } from "@services/app.service";
 import { axGetDataTableList } from "@services/datatable.service";
 import { axGetspeciesGroups } from "@services/observation.service";
-import { axGroupList } from "@services/usergroup.service";
 import { absoluteUrl } from "@utils/basic";
 import React from "react";
