+++ src/pages/document/list.tsx
@@ -1,7 +1,7 @@
 import { DocumentFilterProvider } from "@components/pages/document/common/use-document-filter";
 import DocumentListPageComponent from "@components/pages/document/list";
+import { axGroupList } from "@services/app.service";
 import { axGetListData } from "@services/document.service";
-import { axGroupList } from "@services/usergroup.service";
 import { DEFAULT_FILTER, LIST_PAGINATION_LIMIT } from "@static/documnet-list";
 import { absoluteUrl } from "@utils/basic";
 import React from "react";
