+++ src/pages/index.tsx
@@ -1,5 +1,6 @@
 import HomePageComponent from "@components/pages/home";
-import { axGetGroupHompageDetails, axGroupList } from "@services/usergroup.service";
+import { axGroupList } from "@services/app.service";
+import { axGetGroupHompageDetails } from "@services/usergroup.service";
 import { axGetHomeInfo } from "@services/utility.service";
 import { absoluteUrl } from "@utils/basic";
 import React from "react";
