+++ src/pages/observation/create/single.tsx
@@ -1,9 +1,9 @@
 import { authorizedPageSSP } from "@components/auth/auth-redirect";
 import ObservationCreateSinglePageComponent from "@components/pages/observation/create";
 import { Role } from "@interfaces/custom";
+import { axGroupList } from "@services/app.service";
 import { axGetCreateObservationPageData, axGetspeciesGroups } from "@services/observation.service";
 import { axGetLicenseList } from "@services/resources.service";
-import { axGroupList } from "@services/usergroup.service";
 import { axGetLangList } from "@services/utility.service";
 import { absoluteUrl } from "@utils/basic";
 import React from "react";
