+++ src/utils/userGroup.ts
@@ -1,7 +1,6 @@
 import SITE_CONFIG from "@configs/site-config";
 import { UserGroupIbp } from "@interfaces/observation";
 import { DEFAULT_GROUP } from "@static/constants";
-import { formatDateFromUTC } from "@utils/date";
 import { stringify } from "@utils/query-string";
 
 import { getGroupImage } from "./media";
@@ -38,46 +37,6 @@ export const findCurrentUserGroup = (
 export const getManifestURL = (group: UserGroupIbp) => {
   const { name, icon } = group;
   return `/api/manifest.json?${stringify({ name, icon })}`;
-};
-
-export const formatGroupRules = (rules) => {
-  const groupRules: any[] = [];
-  const {
-    hasUserRule,
-    taxonomicRuleList,
-    spartialRuleList,
-    observedOnDateRule,
-    createdOnDateRuleList
-  } = rules;
-  //populate user rules
-  groupRules.push({ name: "userRule", value: hasUserRule ? "true" : "false" });
-  //populate spatial rules array
-  spartialRuleList?.forEach((item) =>
-    groupRules.push({ id: item.id, name: "spatialRule", value: item.spatialData })
-  );
-
-  //populate taxon rules array
-  taxonomicRuleList?.forEach((item) => {
-    groupRules.push({ id: item.id, name: "taxonomicRule", value: item.taxonomyId });
-  });
-  //observed on date
-  observedOnDateRule?.forEach((item) => {
-    groupRules.push({
-      id: item.id,
-      name: "observedOnDateRule",
-      value: `${formatDateFromUTC(item.fromDate)} to ${formatDateFromUTC(item.toDate)}`
-    });
-  });
-  //created on date
-  createdOnDateRuleList?.forEach((item) => {
-    groupRules.push({
-      id: item.id,
-      name: "createdOnDateRule",
-      value: `${formatDateFromUTC(item.fromDate)} to ${formatDateFromUTC(item.toDate)}`
-    });
-  });
-
-  return groupRules;
 };
 
 export const reorderRemovedGallerySetup = (data, index) => {
